fix(scout-for-lol): satisfy Dare authoring quality gates
Why
Keep the Explore Dare authoring surface within repository lint limits.

What
- extract action executors from the large Dare tool factory
- use replaceAll for target remapping

Verification
- bun run typecheck (packages/scout-for-lol/packages/backend)
- bun x eslint src/explore/dare-tools.ts src/explore/dare-tool-action-executors.ts src/explore/dare-tool-definition-resolution.ts
